fix: improve tunnel administration
